requirements.txt
# ==========================================
# LLM Trading Agent - Python Dependencies
# ==========================================

# Django Framework & REST
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
dj-database-url>=2.1.0

# Authentication (OAuth)
django-allauth>=0.57.0
cryptography>=41.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
pg8000>=1.30.0          # Alternative PostgreSQL adapter

# Web Server
gunicorn>=21.0.0
whitenoise>=6.6.0

# Trading API (Alpaca)
alpaca-py>=0.30.0

# LLM (Google Gemini)
google-genai>=1.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pandas_ta

# Scheduling (for local development)
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
requests>=2.31.0

# Password Hashing
bcrypt>=4.0.0

# ORM (kept for backward compatibility during migration)
sqlalchemy>=2.0.0

# Development tools (optional)
django-extensions>=3.2.0

# ==========================================
# Testing (Future)
# ==========================================
# pytest>=7.4.0
# pytest-django>=4.5.0
# pytest-cov>=4.1.0
